move get_fnames
