+++ dipy/data/__init__.py
@@ -13,7 +13,8 @@ import gzip
 import numpy as np
 from dipy.core.gradients import GradientTable, gradient_table
 from dipy.core.sphere import Sphere, HemiSphere
-from dipy.data.fetcher import (fetch_scil_b0,
+from dipy.data.fetcher import (get_fnames,
+                               fetch_scil_b0,
                                read_scil_b0,
                                fetch_stanford_hardi,
                                read_stanford_hardi,
@@ -194,172 +195,6 @@ def get_sphere(name='symmetric362'):
 
 default_sphere = HemiSphere.from_sphere(get_sphere('repulsion724'))
 small_sphere = HemiSphere.from_sphere(get_sphere('symmetric362'))
-
-
-def get_fnames(name='small_64D'):
-    """Provides full paths to example datasets.
-
-    Parameters
-    ----------
-    name : str
-        the filename/s of which dataset to return, one of:
-        'small_64D' small region of interest nifti,bvecs,bvals 64 directions
-        'small_101D' small region of interest nifti,bvecs,bvals 101 directions
-        'aniso_vox' volume with anisotropic voxel size as Nifti
-        'fornix' 300 tracks in Trackvis format (from Pittsburgh
-            Brain Competition)
-        'gqi_vectors' the scanner wave vectors needed for a GQI acquisitions
-            of 101 directions tested on Siemens 3T Trio
-        'small_25' small ROI (10x8x2) DTI data (b value 2000, 25 directions)
-        'test_piesno' slice of N=8, K=14 diffusion data
-        'reg_c' small 2D image used for validating registration
-        'reg_o' small 2D image used for validation registration
-        'cb_2' two vectorized cingulum bundles
-
-    Returns
-    -------
-    fnames : tuple
-        filenames for dataset
-
-    Examples
-    ----------
-    >>> import numpy as np
-    >>> from dipy.io.image import load_nifti
-    >>> from dipy.data import get_fnames
-    >>> fimg, fbvals, fbvecs = get_fnames('small_101D')
-    >>> bvals=np.loadtxt(fbvals)
-    >>> bvecs=np.loadtxt(fbvecs).T
-    >>> data, affine = load_nifti(fimg)
-    >>> data.shape == (6, 10, 10, 102)
-    True
-    >>> bvals.shape == (102,)
-    True
-    >>> bvecs.shape == (102, 3)
-    True
-
-    """
-    if name == 'small_64D':
-        fbvals = pjoin(DATA_DIR, 'small_64D.bval')
-        fbvecs = pjoin(DATA_DIR, 'small_64D.bvec')
-        fimg = pjoin(DATA_DIR, 'small_64D.nii')
-        return fimg, fbvals, fbvecs
-    if name == '55dir_grad.bvec':
-        return pjoin(DATA_DIR, '55dir_grad.bvec')
-    if name == 'small_101D':
-        fbvals = pjoin(DATA_DIR, 'small_101D.bval')
-        fbvecs = pjoin(DATA_DIR, 'small_101D.bvec')
-        fimg = pjoin(DATA_DIR, 'small_101D.nii.gz')
-        return fimg, fbvals, fbvecs
-    if name == 'aniso_vox':
-        return pjoin(DATA_DIR, 'aniso_vox.nii.gz')
-    if name == 'ascm_test':
-        return pjoin(DATA_DIR, 'ascm_out_test.nii.gz')
-    if name == 'fornix':
-        return pjoin(DATA_DIR, 'tracks300.trk')
-    if name == 'gqi_vectors':
-        return pjoin(DATA_DIR, 'ScannerVectors_GQI101.txt')
-    if name == 'dsi515btable':
-        return pjoin(DATA_DIR, 'dsi515_b_table.txt')
-    if name == 'dsi4169btable':
-        return pjoin(DATA_DIR, 'dsi4169_b_table.txt')
-    if name == 'grad514':
-        return pjoin(DATA_DIR, 'grad_514.txt')
-    if name == "small_25":
-        fbvals = pjoin(DATA_DIR, 'small_25.bval')
-        fbvecs = pjoin(DATA_DIR, 'small_25.bvec')
-        fimg = pjoin(DATA_DIR, 'small_25.nii.gz')
-        return fimg, fbvals, fbvecs
-    if name == 'small_25_streamlines':
-        fstreamlines = pjoin(DATA_DIR, 'EuDX_small_25.trk')
-        return fstreamlines
-    if name == "S0_10":
-        fimg = pjoin(DATA_DIR, 'S0_10slices.nii.gz')
-        return fimg
-    if name == "test_piesno":
-        fimg = pjoin(DATA_DIR, 'test_piesno.nii.gz')
-        return fimg
-    if name == "reg_c":
-        return pjoin(DATA_DIR, 'C.npy')
-    if name == "reg_o":
-        return pjoin(DATA_DIR, 'circle.npy')
-    if name == 'cb_2':
-        return pjoin(DATA_DIR, 'cb_2.npz')
-    if name == "t1_coronal_slice":
-        return pjoin(DATA_DIR, 't1_coronal_slice.npy')
-    if name == 'scil_b0':
-        files, folder = fetch_scil_b0()
-        files = files['datasets_multi-site_all_companies.zip'][2]
-        files = [pjoin(folder, f) for f in files]
-        return [f for f in files if os.path.isfile(f)]
-    if name == 'stanford_hardi':
-        files, folder = fetch_stanford_hardi()
-        fraw = pjoin(folder, 'HARDI150.nii.gz')
-        fbval = pjoin(folder, 'HARDI150.bval')
-        fbvec = pjoin(folder, 'HARDI150.bvec')
-        return fraw, fbval, fbvec
-    if name == 'taiwan_ntu_dsi':
-        files, folder = fetch_taiwan_ntu_dsi()
-        fraw = pjoin(folder, 'DSI203.nii.gz')
-        fbval = pjoin(folder, 'DSI203.bval')
-        fbvec = pjoin(folder, 'DSI203.bvec')
-        return fraw, fbval, fbvec
-    if name == 'sherbrooke_3shell':
-        files, folder = fetch_sherbrooke_3shell()
-        fraw = pjoin(folder, 'HARDI193.nii.gz')
-        fbval = pjoin(folder, 'HARDI193.bval')
-        fbvec = pjoin(folder, 'HARDI193.bvec')
-        return fraw, fbval, fbvec
-    if name == 'isbi2013_2shell':
-        files, folder = fetch_isbi2013_2shell()
-        fraw = pjoin(folder, 'phantom64.nii.gz')
-        fbval = pjoin(folder, 'phantom64.bval')
-        fbvec = pjoin(folder, 'phantom64.bvec')
-        return fraw, fbval, fbvec
-    if name == 'stanford_labels':
-        files, folder = fetch_stanford_labels()
-        return pjoin(folder, "aparc-reduced.nii.gz")
-    if name == 'syn_data':
-        files, folder = fetch_syn_data()
-        t1_name = pjoin(folder, 't1.nii.gz')
-        b0_name = pjoin(folder, 'b0.nii.gz')
-        return t1_name, b0_name
-    if name == 'stanford_t1':
-        files, folder = fetch_stanford_t1()
-        return pjoin(folder, 't1.nii.gz')
-    if name == 'stanford_pve_maps':
-        files, folder = fetch_stanford_pve_maps()
-        f_pve_csf = pjoin(folder, 'pve_csf.nii.gz')
-        f_pve_gm = pjoin(folder, 'pve_gm.nii.gz')
-        f_pve_wm = pjoin(folder, 'pve_wm.nii.gz')
-        return f_pve_csf, f_pve_gm, f_pve_wm
-    if name == 'ivim':
-        files, folder = fetch_ivim()
-        fraw = pjoin(folder, 'ivim.nii.gz')
-        fbval = pjoin(folder, 'ivim.bval')
-        fbvec = pjoin(folder, 'ivim.bvec')
-        return fraw, fbval, fbvec
-    if name == 'tissue_data':
-        files, folder = fetch_tissue_data()
-        t1_name = pjoin(folder, 'tissue_data', 't1_brain.nii.gz')
-        t1d_name = pjoin(folder, 'tissue_data', 't1_brain_denoised.nii.gz')
-        ap_name = pjoin(folder, 'tissue_data', 'power_map.nii.gz')
-        return t1_name, t1d_name, ap_name
-    if name == 'cfin_multib':
-        files, folder = fetch_cfin_multib()
-        t1_name = pjoin(folder, 'T1.nii')
-        fraw = pjoin(folder, '__DTI_AX_ep2d_2_5_iso_33d_20141015095334_4.nii')
-        fbval = pjoin(folder,
-                      '__DTI_AX_ep2d_2_5_iso_33d_20141015095334_4.bval')
-        fbvec = pjoin(folder,
-                      '__DTI_AX_ep2d_2_5_iso_33d_20141015095334_4.bvec')
-        return fraw, fbval, fbvec, t1_name
-    if name == 'target_tractrogram_hcp':
-        files, folder = fetch_target_tractogram_hcp()
-        return pjoin(folder, 'target_tractogram_hcp', 'hcp_tractogram',
-                     'streamlines.trk')
-    if name == 'bundle_atlas_hcp842':
-        files, folder = fetch_bundle_atlas_hcp842()
-        return get_bundle_atlas_hcp842()
 
 
 def _gradient_from_file(filename):
